spark/requirements.txt
kafka-python
delta-spark